+++ apps/backend/src/config/migrations.ts
@@ -36,6 +36,7 @@ import { RenameDonationMatchingStatus1771260403657 } from '../migrations/1771260
 import { DropDonationTotalColumns1772241115031 } from '../migrations/1772241115031-DropDonationTotalColumns';
 import { FixTrackingLinks1773041840374 } from '../migrations/1773041840374-FixTrackingLinks';
 import { CleanupRequestsAndAllocations1771821377918 } from '../migrations/1771821377918-CleanupRequestsAndAllocations';
+import { MakeFoodRescueRequired1773889925002 } from '../migrations/1773889925002-MakeFoodRescueRequired.ts';
 
 const schemaMigrations = [
   User1725726359198,
@@ -76,6 +77,7 @@ const schemaMigrations = [
   DropDonationTotalColumns1772241115031,
   FixTrackingLinks1773041840374,
   CleanupRequestsAndAllocations1771821377918,
+  MakeFoodRescueRequired1773889925002,
 ];
 
 export default schemaMigrations;
