+++ apps/backend/src/donationItems/donationItems.service.ts
@@ -52,15 +52,4 @@ export class DonationItemsService {
 
     return this.repo.save(donationItem);
   }
-
-  async updateDonationItemQuantity(itemId: number): Promise<DonationItem> {
-    validateId(itemId, 'Donation Item');
-
-    const donationItem = await this.repo.findOneBy({ itemId });
-    if (!donationItem) {
-      throw new NotFoundException(`Donation item ${itemId} not found`);
-    }
-    donationItem.quantity -= 1;
-    return this.repo.save(donationItem);
-  }
 }
